Install mandoc manually in travis; update Makefile
It's not in Trusty, nor trusty backports, and Sid's is too new. Get grab an old
version and use it locally.

Mandoc's flags have changed a bit, and they no longer ship example.css.

Much to my embarrassment, my Makefile knowledge is so poor that I can manage to
figure out how to conditionally declare a MANDOC variable depending on whether
it's already int he path or not. Sigh...